# Speed up index loading from SQLite for SeqIO, SearchIO, etc