--- a/CONTRIB.rst
+++ b/CONTRIB.rst
@@ -202,6 +202,7 @@ please open an issue on GitHub or mention it on the mailing list.
 - Paul T. Bathen
 - Peter Bienstman <Peter.Bienstman at domain rug.ac.be>
 - Peter Cock <https://github.com/peterjc>
+- Peter Kerpedjiev <https://github.com/pkerpedjiev>
 - Peter Slickers <piet at domain clondiag.com>
 - Philip Bergstrom <https://github.com/phber>
 - Phillip Garland <pgarland at gmail>