The current `SELECT COUNT(key) FROM offset_data;` call to obtain the number of sequences in the database is slow for substantial fasta files.

Because the index is read-only, this can be replaced with `SELECT MAX(_ROWID_) FROM offset_data;` for orders of magnitude better performance.

For example, on a fasta with 17M sequences, loading the index with the current `COUNT(key)` call takes about 4 seconds. With `MAX(_ROWID_)` it takes 2.64ms.

On a fasta file with 307M sequences, loading the index with `COUNT(key)` takes 1 minute and 20 seconds. With `MAX(_ROWID_)` it takes 10ms. 

<!--- Please read each of the following items and confirm by replacing
 !--the [ ] with a [X] --->

- [x] I hereby agree to dual licence this and any previous contributions under both
the _Biopython License Agreement_ **AND** the _BSD 3-Clause License_.

- [x] I have read the ``CONTRIBUTING.rst`` file, have run ``flake8`` locally, and
understand that AppVeyor and TravisCI will be used to confirm the Biopython unit
tests and style checks pass with these changes.

- [x] I have added my name to the alphabetical contributors listings in the files
``NEWS.rst`` and ``CONTRIB.rst`` as part of this pull request, am listed
already, or do not wish to be listed. (*This acknowledgement is optional.*)