@peterjc 2019-06-17
Discussion on eg https://stackoverflow.com/questions/8988915/sqlite-count-slow-on-big-tables suggests this is a known SQLite weakness, and common work around.

The only reason we're doing this check is for data consistency, so I think your change is fine - even given the possibility of other tools/scripts building or modifying our SQLite index format.

However, I do think the code change does need some explanation as a comment, e.g.

```python
# Using "SELECT MAX(key) FROM offset_data;" is too slow
```

@bow Do you agree?

@pkerpedjiev Do you want to put your name in ``NEWS.rst`` too?