@pkerpedjiev 2019-06-17
@peterjc Thanks for the quick response! I added a comment with a link to the StackOverflow question and a short explanation and also added my name to NEWS.rst.